Addd connection between tables, added task dto  mapper

+++ src/main/java/com/example/todolist/services/user/UserService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface UserService {
     List<User> getAllUsers();
     void saveUser(User user);
-    User getUser(int id);
-    void deleteUser(int id);
+    User getUser(Long id);
+    void deleteUser(Long id);
     void deleteAllUsers();
     List<User> findAllUsers();
 }
